requirements.txt
requests
requests_oauthlib
apscheduler
python-dotenv
flask